+++ lms/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class LmsConfig(AppConfig):
+    name = 'learning management system'
