+++ nodes/admin.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+from nodes.models import *
+
+
+admin.site.register(Subject)
+admin.site.register(Module)
+admin.site.register(Unit)
+admin.site.register(Lesson)
+admin.site.register(LessonResult)
