--- teaching/apps.py
@@ -1,7 +1,0 @@
-# coding=utf-8
-from django.apps import AppConfig
-
-
-class AppConfig(AppConfig):
-    name = 'teaching'
-    verbose_name = 'Обучение'
