+++ templates/components/modals.html
@@ -9,7 +9,7 @@
                 </button>
                 <h4 class="modal-title">Авторизация</h4>
                 <a href="{% url 'reg_view' %}">Я не
-                        зарегистрирован</a>
+                    зарегистрирован</a>
             </div>
             <div class="modal-body">
                 <div class="row">
@@ -33,13 +33,10 @@
                     </div>
                     <div class="col-xs-6">
                         {% url 'login_view' as login_url %}
-                        <a class="btn btn-primary btn-block" href="{% vkontakte_auth_link login_url %}">VK</a>
+                        <a class="btn btn-primary btn-block" href="{% vk_auth_link login_url %}">VK</a>
                     </div>
                 </div>
             </div>
         </div>
     </div>
-    <!-- /.modal-content -->
 </div>
-<!-- /.modal-dialog -->
-</div><!-- /.modal -->
