+++ templates/login.html
@@ -11,7 +11,7 @@
                             <br/>
                             <h2>Авторизация</h2>
                             <p>
-                                <a href="#" data-dismiss="modal" data-toggle="modal" data-target="#regModal">Я не зарегистрирован</a>
+                                <a href="{% url 'reg_view' %}">Я не зарегистрирован</a>
                             </p>
                         </div>
                     </div>
@@ -36,7 +36,7 @@
                             </div>
                             <div class="col-xs-6">
                                 {% url 'login_view' as login_url %}
-                                <a class="btn btn-primary btn-block" href="{% vkontakte_auth_link login_url %}">VK</a>
+                                <a class="btn btn-primary btn-block" href="{% vk_auth_link login_url %}">VK</a>
                             </div>
                         </div>
                     </div>
