+++ templates/teaching/cabinet.html
@@ -8,18 +8,18 @@
             <li class="breadcrumb-item active">Кабинет учителя</li>
         </ol>
         <div id="accordion" aria-multiselectable="true">
-            {% for studentteacher in studentteachers %}
+            {% for student_teacher in student_teachers %}
                 <div class="card">
                     <div class="card-header">
-                        <a href="{% url 'user_view' studentteacher.student.pk %}">
-                            <span class="lead">{{ studentteacher.student }}</span>
+                        <a href="{% url 'user_view' student_teacher.student.pk %}">
+                            <span class="lead">{{ student_teacher.student }}</span>
                         </a>
-                        <a data-toggle="collapse" href="#collapse{{ studentteacher.pk }}" aria-expanded="true" aria-controls="collapse{{ studentteacher.pk }}">
+                        <a data-toggle="collapse" href="#collapse{{ student_teacher.pk }}" aria-expanded="true" aria-controls="collapse{{ student_teacher.pk }}">
                             Открыть
                         </a>
                     </div>
-                    <div id="collapse{{ studentteacher.pk }}" class="card-collapse collapse card-block" role="tabpanel">
-                        {% if studentteacher.tasks.count > 0 %}
+                    <div id="collapse{{ student_teacher.pk }}" class="card-collapse collapse card-block" role="tabpanel">
+                        {% if student_teacher.tasks.count > 0 %}
                             <p class="lead">Задания</p>
                             <table class="table table-bordered">
                                 <thead>
@@ -30,11 +30,11 @@
                                     </tr>
                                 </thead>
                                 <tbody>
-                                {% for task in studentteacher.tasks %}
+                                {% for task in student_teacher.tasks %}
                                     <tr>
                                         <td><a href="{% url 'lesson_view' task.lesson.id %}" target="_blank">{{ task.lesson.title }}</a></td>
                                         <td>{{ task.datetime_to }}</td>
-                                        <td>{% if task.is_finished %}ВЫПОЛНИЛ {{ task.studentlesson.score }}/{{ task.studentlesson.max_score }} баллов{% else %}не сделал{% endif %}</td>
+                                        <td>{% if task.is_finished %}ВЫПОЛНИЛ {{ task.lesson_result.score }}/{{ task.lesson_result.max_score }} баллов{% else %}не сделал{% endif %}</td>
                                     </tr>
                                 {% endfor %}
                                 </tbody>
