--- templates/teaching/floatquestion.html
@@ -1,74 +1,0 @@
-{% extends 'teaching/block_base.html' %}
-
-{% block block-content %}
-
-    <p class="lead">Задача #{{ floatquestion.id }}</p>
-    <p>{{ floatquestion.rendered_question_text|safe }}</p>
-    {% if floatquestion.image %}
-        <p class="text-xs-center">
-            <img src="{{ floatquestion.image.url }}">
-        </p>
-    {% endif %}
-    <hr/>
-    <div class="row">
-        <div class="col-md-4">
-            <input id="answerfield" class="form-control" type="text" name="answer" placeholder="Введите ответ">
-        </div>
-    </div>
-    <hr>
-    <div id="buttons">
-        <button id="answer-action" class="btn btn-primary" disabled>Проверить</button>
-    </div>
-
-{% endblock %}
-
-{% block block-scripts %}
-
-    <script>
-        {% if lesson %}
-            $(document).ready(function() {
-                $('#answer-action').click(function() {
-                    var answer = $("#answerfield").val();
-                    answer = answer.replace(/,/g, '.');
-
-                    {% if last_block %}
-                        var tpl = '<a href="{% url 'lesson_final_view' lesson.id %}" class="btn btn-primary btn-lg">Закончить урок</a>';
-                    {% elif next_block_num %}
-                        var tpl = '<a href="{% url 'lesson_block_view' lesson.id next_block_num %}" class="btn btn-primary btn-lg">Следующий блок ></a>';
-                    {% endif %}
-
-                    $.ajax({
-                        url: '{% url "lesson_block_view" lesson.pk block_num %}',
-                        data: {
-                            csrfmiddlewaretoken: '{{ csrf_token }}',
-                            action: 'answer',
-                            answer: answer
-                        },
-                        async: true,
-                        success: function (responseData, textStatus) {
-                            tpl = '<p class="text-muted">' + responseData + '</p>' + tpl
-                            $('#buttons').fadeOut('slow', function () {
-                                $('#buttons').replaceWith(tpl);
-                            });
-                        },
-                        error: function (response, status, errorThrown) {
-                            alert('Все плохо, расскажите нам про эту ошибку \n\r\n\r' + response + status + errorThrown);
-                        },
-                        type: "POST",
-                        dataType: "text"
-                    });
-                });
-            });
-
-            $(document).on('input', '#answerfield', function () {
-                var len = $("#answerfield").val().length;
-                if (len > 0) {
-                    $("#answer-action").attr('disabled', false);
-                } else {
-                    $("#answer-action").attr('disabled', true);
-                }
-            });
-        {% endif %}
-    </script>
-
-{% endblock %}
