+++ templates/teaching/text_block.html
@@ -1,0 +1,52 @@
+{% extends 'teaching/block_base.html' %}
+{% load django_markdown %}
+{% load globaltags %}
+
+{% block block-content %}
+
+    <h1 class="hidden-sm-down">{{ text_block.title }}</h1>
+    <h1 class="hidden-sm-up h2">{{ text_block.title }}</h1>
+    <hr>
+    <div class="paragraph-body">{{ text_block.body|markdown_rendered|safe }}</div>
+    <hr>
+    <div id="buttons">
+        <button id="read-action" class="btn btn-primary">Я прочитал</button>
+    </div>
+
+{% endblock %}
+
+{% block block-scripts %}
+
+    <script>
+        {% if lesson %}
+            $(document).ready(function() {
+                $('#read-action').click(function() {
+
+                    {% if last_block %}
+                        var url = "{% url 'lesson_final_view' lesson.id %}";
+                    {% elif next_block_num %}
+                        var url = "{% url 'lesson_block_view' lesson.id next_block_num %}";
+                    {% endif %}
+
+                    $.ajax({
+                        url: '{% url "lesson_block_view" lesson.pk block_num %}',
+                        data: {
+                            csrfmiddlewaretoken: '{{ csrf_token }}',
+                            action: 'read'
+                        },
+                        async: true,
+                        success: function (responseData, textStatus) {
+                            window.location.href = url;
+                        },
+                        error: function (response, status, errorThrown) {
+                            alert('Все плохо, расскажите нам про эту ошибку \n\r\n\r' + response + status + errorThrown);
+                        },
+                        type: "POST",
+                        dataType: "text"
+                    });
+                });
+            });
+        {% endif %}
+    </script>
+
+{% endblock %}
