+++ templates/user.html
@@ -23,7 +23,7 @@
             <br>
             <br>
             <p>
-                <a href="{% vkontakte_profile_link user.vkuserid %}" target="_blank">
+                <a href="{% vk_profile_link user.vkuserid %}" target="_blank">
                     <img src="{% static 'images/vk.png' %}" width="32">
                 </a>
             </p>
