+++ utils/context_processors.py
@@ -1,29 +1,23 @@
-from users.models import User
-from teaching.models import Task
+from tasks.models import Task
 
 
-def loggedin_user(request):
+def logged_in_user(request):
     context = {'is_logged_in': request.user.is_authenticated()}
-    try:
-        if request.user.is_authenticated():
-            user = User.objects.get(email=request.user.email)
-            context['current_user'] = user
-        else:
-            context['current_user'] = None
-    except User.DoesNotExist:
-        pass
+    if request.user.is_authenticated():
+        context['current_user'] = request.user
+    else:
+        context['current_user'] = None
     return context
 
 
 def tasks_counter(request):
     context = {}
-    try:
-        if request.user.is_authenticated():
-            user = User.objects.get(email=request.user.email)
+    if request.user.is_authenticated():
+        try:
             tasks = Task.objects.filter(student=request.user, is_finished=False)
             context['glob_tasks'] = tasks
-        else:
+        except Task.DoesNotExist:
             context['glob_tasks'] = None
-    except User.DoesNotExist:
-        pass
+    else:
+        context['glob_tasks'] = None
     return context
